fixed most bugs with program and finished up functions. timing for search hashmap still sketchy

+++ Code/odom.h
@@ -5,13 +5,13 @@ class Odom
     double px, py, pz;
     double lvx, lvy, lvz;
     double avx, avy, avz;
-    double t;
+    int t;
 
 public:
 
     Odom(double px, double py, double pz,
         double lvx, double lvy, double lvz,
-        double avx, double avy, double avz, double t) :
+        double avx, double avy, double avz, int t) :
         px(px), py(py), pz(pz), lvx(lvx), lvy(lvy), lvz(lvz), avx(avx), avy(avy), avz(avz), t(t) {}
 
     double px_() const { return px; }
